+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.16.0
 
 ### Enhancements
+- **Export Changes**
+  - Exports now include separate folders with a `README` for each export option
 - **Gem Changes**
   - Updated to pg 0.15.1
   - Updated to jbuilder 1.3.0